SIGPLAN: revert changes to \thesection and \thesubsection
The change to command `\thesection` to include a trailing period was
intended to only effect the section number appearing in a section
heading.  However `\thesection` is captured by `\label` (and so the
trailing period erroneously in the corresponding `\ref`) and used in
theorem environments that derive their number from the current
section (an so the trailing period erroneously leads to double periods:
"Theorem 4..1").