Clean up of `timestamp` code.
The original code used `\newcount` instead of `\newcounter` so we got the
right rounding behavior.  (See the comment in the original code.)

I've realized how to get the right rounding from `\newcounter`, which slightly
simplifies the code and is implemented in this commit.  This should not change
any functionality or behavior.