Fix sorting of `manual` BibTeX entries.
Manual entries that have an `editor` but no `author` were not being sorted
correctly because in `presort`, `author.organization.sort` was being called,
which doesn't look at the `editor`.  A similar thing was happening in
`calc.label`.  The fix was to add and use an `author.editor.organization.sort`
function (and similarly in `calc.label` with the `.label` functions).

Note: The diff on this is wonky.  `author.organization.sort` was removed (and
replaced by `author.editor.organization.sort`).  It was not renamed to
`editor.organization.sort` like the diff suggests.