Remove spurious "numpages field, but no articleno field" warning
The containing function, `format.page.count`, didn't actually check if the
`numpages` field is empty.  That function was called from two places.  The
first is in `format.articleno.numpages`, which already checks and issues
warnings for this.  The second is in `calc.format.page.count` and doesn't
check if `numpages` is empty.  Thus this warning could be issued on entries
for which there was no `numpages` field.

Since the first place already checks and warns and the second place isn't
related to anything having to do with `articleno` the fix is to just remove
the warning.

This commit also reformats a couple warnings to be consistent with the
warnings in the rest of the file.