Remove duplicated year from header
The style guide shows that the `short name` of a conference should include the
year, such as:

    \acmConference[SenSys'17]{The 15th ACM Conference on Embedded Networked Sensor Systems}{November 6-8, 2017}{Delft, The Netherlands}

Prior to this patch, that creates column headers that look like:

    SenSys'17, November 6-8, 2017, Delft, The Netherlands

Which is visually confusing and redundant to have both `'17` and 2017.
It is also simply a lot of text at the top of each column.

After this patch, the header simply reads:

    SenSys'17, Delft, The Netherlands