Set PDF keywords to keywords, and PDF subject to CCS concepts
Previously the PDF keywords were set to the CCS concepts, and the
paper's visible list of keywords appeared nowhere at all in the PDF
metadata.  If we want both keywords and CCS concepts to be reflected
in the PDF metadata, then clearly PDF keywords are where the visible
keywords belong.  That leaves the PDF subject available, which seems a
suitable place to park the CCS concepts text.