Use \urldef to expand \url before \showURL/\showDOI for all URLs.
This makes URLs with characters that are normally special to TeX,
such as %, #, and _, work correctly in all cases.  (\url reads its
argument with special catcodes, but it doesn't get a chance to do
that if \showURL or \showDOI is expanded first.)

This necessitated removal of the "semicolon-separated list of URLs"
feature, but ; can and does appear inside URLs, so that feature was
highly questionable anyway.