implement a 'nonacm' option to produce standalone (non-ACM) documents
The 'nonacm' option is meant to be used for documents that were
prepared for an ACM submission using the acmart class and one of the
formats, but are not accepted for publication. They wish to make the
document available on their webpage, circulate it to colleagues, but
keeping the acmart documentclass includes a lot of
ACM-publication-specific content that is irrelevant to
them.

Currently their only choice, if they want a version of their work that
displays nicely, is to change their documentclass to another LaTeX
class, which is tedious as it involves reflows etc., but also
rewriting the author, affiliation, institute information to use
different macros. They may later decide to resubmit the work to
another ACM venue, and have to change this all over again.

The new 'nonacm' option disables the ACM-venues-specific content in
the acmart output (name of journal or conference, copyright
statements, DOI...) to produce a standalone, independent document that
authors can release.
(The 'manuscript' or 'authorversion' are simliar but do not solve this
use-case: this is not an "author version of an ACM publication" or
a "manuscript submitted to the ACM".)

This option is complementary to the choice of a format option
(acmsmall, acmlarge...), it is not an extra format. Setting 'nonacm'
should respect the existing format (although of course the page breaks
may be placed differently). For example, most headers and footers set
by the format are preserved, only the ones that reference a specific
journal/conference are omitted.

By default, the 'nonacm' option sets 'printacmref=false', so the "ACM
Reference Format" paragraph is not included. Users can explicitly
enable '\settopmatter{printacmref=true}' to re-enable it.