doc: change \acmBadgeR to \acmBadge
I find it confusing that the documentation of \acmBadge
has an example that uses \acmBadgeR, not \acmBadge.
After looking at the history, I found the commit b17e331a, which
replaces \acmBadgeR with \acmBadge (though internally it is just
a syntactic sugar of \acmBadgeR). However, the example in the
documentation was not similarly updated to reflect this.
This PR fixes the issue.